requirements.txt
amqp==5.1.1
APScheduler==3.9.1.post1
billiard==3.6.4.0
celery==5.2.7
certifi
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.2.2
idna==2.8
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.1.1
prompt-toolkit==3.0.33
PyJWT==2.6.0
pymongo==3.13.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlmap==1.6.10
tzdata==2022.7
tzlocal==4.2
urllib3==1.24.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.10.0
requests
logger
pymongo
ConfigParser
pyjwt
flask
sqlmapamqp==5.1.1
APScheduler==3.9.1.post1
billiard==3.6.4.0
celery==5.2.7
certifi
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.2.2
idna==2.8
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.0.3
kombu==5.2.4
MarkupSafe==2.1.1
prompt-toolkit==3.0.33
PyJWT==2.6.0
pymongo==3.13.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
sqlmap==1.6.10
tzdata==2022.7
tzlocal==4.2
urllib3==1.24.3
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.10.0
Flask==2.2.2
Jinja2==3.0.3
PyJWT==2.6.0
pymongo==3.13.0
requests==2.28.1
sqlmap==1.7
apscheduler
